+++ lib/cli.rs
@@ -207,6 +207,7 @@ pub async fn transfer(
     transfer_id: &str,
     deploy_params: DeployStrParams<'_>,
     payment_params: PaymentStrParams<'_>,
+    session_args: Vec<&str>,
 ) -> Result<SuccessResponse<PutDeployResult>, CliError> {
     let rpc_id = parse::rpc_id(maybe_rpc_id);
     let verbosity = parse::verbosity(verbosity_level);
@@ -218,6 +219,7 @@ pub async fn transfer(
         deploy_params,
         payment_params,
         false,
+        session_args,
     )?;
     crate::put_deploy(rpc_id, node_address, verbosity, deploy)
         .await
@@ -246,6 +248,7 @@ pub async fn speculative_transfer(
     transfer_id: &str,
     deploy_params: DeployStrParams<'_>,
     payment_params: PaymentStrParams<'_>,
+    session_args: Vec<&str>,
 ) -> Result<SuccessResponse<SpeculativeExecResult>, CliError> {
     let rpc_id = parse::rpc_id(maybe_rpc_id);
     let verbosity = parse::verbosity(verbosity_level);
@@ -257,6 +260,7 @@ pub async fn speculative_transfer(
         deploy_params,
         payment_params,
         false,
+        session_args,
     )?;
     let speculative_exec = parse::block_identifier(maybe_block_id)?;
     crate::speculative_exec(rpc_id, node_address, speculative_exec, verbosity, deploy)
@@ -281,6 +285,7 @@ pub fn make_transfer(
     deploy_params: DeployStrParams<'_>,
     payment_params: PaymentStrParams<'_>,
     force: bool,
+    session_args: Vec<&str>,
 ) -> Result<(), CliError> {
     let output = parse::output_kind(maybe_output_path, force);
     let deploy = deploy::new_transfer(
@@ -291,6 +296,7 @@ pub fn make_transfer(
         deploy_params,
         payment_params,
         true,
+        session_args,
     )?;
     crate::output_deploy(output, &deploy).map_err(CliError::from)
 }