+++ src/deploy/make_transfer.rs
@@ -27,11 +27,16 @@ impl ClientCommand for MakeTransfer {
             .arg(common::force::arg(
                 creation_common::DisplayOrder::Force as usize,
                 true,
-            ));
+            ))
+            .arg(creation_common::arg_simple::session::arg())
+            .arg(creation_common::args_json::session::arg())
+            .arg(creation_common::args_complex::session::arg());
+
         let subcommand = creation_common::apply_common_payment_options(
             subcommand,
             Some(common::DEFAULT_TRANSFER_PAYMENT_AMOUNT),
         );
+
         creation_common::apply_common_creation_options(subcommand, false, false)
     }
 
@@ -43,6 +48,7 @@ impl ClientCommand for MakeTransfer {
         let target_account = transfer::target_account::get(matches);
         let transfer_id = transfer::transfer_id::get(matches);
 
+        let str_params = creation_common::arg_simple::session::get(matches);
         let secret_key = common::secret_key::get(matches).unwrap_or_default();
         let timestamp = creation_common::timestamp::get(matches);
         let ttl = creation_common::ttl::get(matches);
@@ -68,6 +74,7 @@ impl ClientCommand for MakeTransfer {
             },
             payment_str_params,
             force,
+            str_params,
         )
         .map(|_| {
             Success::Output(if maybe_output_path.is_empty() {