+++ src/deploy/transfer.rs
@@ -11,14 +11,12 @@ pub(super) mod amount {
     use super::*;
 
     const ARG_NAME: &str = "amount";
-    const ARG_SHORT: char = 'a';
     const ARG_VALUE_NAME: &str = "512-BIT INTEGER";
     const ARG_HELP: &str = "The number of motes to transfer";
 
     pub(in crate::deploy) fn arg() -> Arg {
         Arg::new(ARG_NAME)
             .long(ARG_NAME)
-            .short(ARG_SHORT)
             .required_unless_present(creation_common::show_simple_arg_examples::ARG_NAME)
             .required_unless_present(creation_common::show_json_args_examples::ARG_NAME)
             .value_name(ARG_VALUE_NAME)
@@ -126,6 +124,7 @@ impl ClientCommand for Transfer {
         let target_account = target_account::get(matches);
         let transfer_id = transfer_id::get(matches);
 
+        let str_params = creation_common::arg_simple::session::get(matches);
         let maybe_rpc_id = common::rpc_id::get(matches);
         let node_address = common::node_address::get(matches);
         let verbosity_level = common::verbose::get(matches);
@@ -156,6 +155,7 @@ impl ClientCommand for Transfer {
                     session_account: &session_account,
                 },
                 payment_str_params,
+                str_params,
             )
             .await
             .map(Success::from)
@@ -175,6 +175,7 @@ impl ClientCommand for Transfer {
                     session_account: &session_account,
                 },
                 payment_str_params,
+                str_params,
             )
             .await
             .map(Success::from)